UIViewControllerRepresentable Coordinator and Artwork Display - Added Coordinator to UIViewControllerRepresentable for MediaPickerController - Moved MPMediaPickerControllerDelegate to Coordinator - MPMusicPlayerControllerQueueDidChange notification to publish artwork (using Combine/NotificationCenter publisher)
