+++ MELMK/MediaPickerController.swift
@@ -5,7 +5,6 @@
 //  Created by Zamel Rakim on 1/11/21.
 //
 
-
 import SwiftUI
 import UIKit
 import MediaPlayer
@@ -15,10 +14,14 @@ struct MediaPickerController : UIViewControllerRepresentable {
     
     var delegateObj: AMAPI?
     
+    func makeCoordinator() -> Coordinator {
+        Coordinator(self)
+    }
+    
     func makeUIViewController(context: Context) -> MPMediaPickerController {
         let controller = MPMediaPickerController(mediaTypes: .music)
         controller.allowsPickingMultipleItems = true
-        controller.delegate = self.delegateObj
+        controller.delegate = context.coordinator
         return controller
     }
     
@@ -26,4 +29,27 @@ struct MediaPickerController : UIViewControllerRepresentable {
         return
     }
     
+    class Coordinator: NSObject, MPMediaPickerControllerDelegate {
+        var parent: MediaPickerController
+        var player: MPMusicPlayerApplicationController?
+        
+        init(_ mediaPickerController: MediaPickerController) {
+            parent = mediaPickerController
+            player = parent.delegateObj?.player
+        }
+        
+        func mediaPicker(_ mediaPicker: MPMediaPickerController, didPickMediaItems mediaItemCollection: MPMediaItemCollection) {
+            print("Choose A Song(s)")
+            
+            player?.setQueue(with: mediaItemCollection)
+            mediaPicker.dismiss(animated: true)
+            player?.prepareToPlay()
+            player?.play()
+        }
+        
+        func mediaPickerDidCancel(_ mediaPicker: MPMediaPickerController) {
+            print("Canceled Media Picker")
+            mediaPicker.dismiss(animated: true)
+        }
+    }
 }
